+++ Engine/Components/CameraController.cs
@@ -23,35 +23,35 @@ namespace SourceRewrite.Components
             camera = GameObject.GetComponentFromType<Camera>();
         }
 
-        public override void Update(double deltaTime)
+        public override void Update(float deltaTime)
         {
             // Forward Movement
             if (Input.GetKeyDown(Key.W))
             {
-                GameObject.Transform.Position = GameObject.Transform.Position + MovementSpeed * (float)deltaTime * GameObject.Transform.Forward;
+                GameObject.Transform.Position = GameObject.Transform.Position + MovementSpeed * deltaTime * GameObject.Transform.Forward;
             }
 
             // Backward Movement
             if (Input.GetKeyDown(Key.S))
             {
-                GameObject.Transform.Position = GameObject.Transform.Position - MovementSpeed * (float)deltaTime * GameObject.Transform.Forward;
+                GameObject.Transform.Position = GameObject.Transform.Position - MovementSpeed * deltaTime * GameObject.Transform.Forward;
             }
 
             // Left Movement
             if (Input.GetKeyDown(Key.A))
             {
-                GameObject.Transform.Position = GameObject.Transform.Position - MovementSpeed * (float)deltaTime * GameObject.Transform.Right;
+                GameObject.Transform.Position = GameObject.Transform.Position - MovementSpeed * deltaTime * GameObject.Transform.Right;
             }
 
             // Right Movement
             if (Input.GetKeyDown(Key.D))
             {
-                GameObject.Transform.Position = GameObject.Transform.Position + MovementSpeed * (float)deltaTime * GameObject.Transform.Right;
+                GameObject.Transform.Position = GameObject.Transform.Position + MovementSpeed * deltaTime * GameObject.Transform.Right;
             }
 
             // Mouse Movement
-            float mouseX = -Input.GetMouseXMovement() * Sensitivity * (float)deltaTime;
-            float mouseY = -Input.GetMouseYMovement() * Sensitivity * (float)deltaTime;
+            float mouseX = -Input.GetMouseXMovement() * Sensitivity * deltaTime;
+            float mouseY = -Input.GetMouseYMovement() * Sensitivity * deltaTime;
 
             GameObject.Transform.RotateBy(mouseY, mouseX, 0);
         }
